4.2.4
added EndgameMarker
added collision related to it
added code for tileresources to parse it
added init in game map

+++ PhysicEngine/Player/Rect.h
@@ -50,7 +50,7 @@ class Rect : public GameShape  {
 	sf::Texture texture;
 	sf::Sprite sprite;
 	float DT;
-	float sizeOfSprite;
+	int sizeOfSprite;
 	//int points;
 	bool isLarge;
 	sf::Vector2f airResistanceConstants;
